pyproject.toml
[project]
name = "chatbot-whatsapp"
version = "1.0.0"
description = "Chatbot inteligente para WhatsApp con LangGraph - Migrado desde n8n"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Marlon Pernia", email = "marlon@example.com" }
]

dependencies = [
    # LangGraph & LangChain
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # Redis
    "redis>=5.0.0",
    
    # OpenAI
    "openai>=1.12.0",
    
    # Google
    "google-generativeai>=0.5.0",
    
    # HTTP Client
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.9",

    # RAG & Postgres
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.5",

    # Document extraction
    "pdfminer.six>=20231228",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "chardet>=5.2.0",
    
    # Config & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    
    # Logging
    "structlog>=24.1.0",

    # Observability (Phoenix + OpenTelemetry)
    "arize-phoenix-otel>=0.14.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "openinference-instrumentation-langchain>=0.1.0",
    "openinference-instrumentation-openai>=0.1.0",
    
    # Email (optional)
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
